Add proptypes
