+++ src/components/HomeView/HomeView.js
@@ -1,17 +1,22 @@
 import './HomeView.css'
-import Header from '../Header/Header';
 import SearchDrink from '../SearchDrink/SearchDrink';
 import DisplayDrinks from '../DisplayDrinks/DisplayDrinks';
+import PropTypes from 'prop-types'
 
-const HomeView = ({ setSearch, fetchAlcohol, drinks, addFavoriteDrinks }) => {
+const HomeView = ({fetchAlcohol, drinks, addFavoriteDrinks }) => {
 
   return (
     <div className='home-view'>
-      {/* <Header /> */}
-      <SearchDrink setSearch={setSearch} fetchAlcohol={fetchAlcohol}/>
+      <SearchDrink fetchAlcohol={fetchAlcohol}/>
       <DisplayDrinks drinks={drinks} addFavoriteDrinks={addFavoriteDrinks}/>
     </div>
   ) 
 }
 
 export default HomeView;
+
+HomeView.propTypes = {
+  fetchAlcohol: PropTypes.func.isRequired,
+  drinks: PropTypes.array.isRequired,
+  addFavoriteDrinks: PropTypes.func.isRequired
+}
